+++ beecell/cache/client.py
@@ -3,30 +3,30 @@
 # (C) Copyright 2018-2023 CSI-Piemonte
 
 import logging
+import ujson as json
 import pickle
 import codecs
 from typing import Any
-import ujson as json
-import redis
+from beecell.db.manager import RedisManager
 from beecell.simple import truncate, jsonDumps
 
 
 class CacheClient(object):
     """ """
 
-    def __init__(self, redis: redis.StrictRedis, prefix="cache."):
+    def __init__(self, redis_manager: RedisManager, prefix="cache."):
         """Initialize cache client
 
-        :param redis: redis manager reference (redis.StrictRedis or StrictRedisCluster instance)
+        :param redis: redis manager reference (RedisManager)
         :param prefix: chache key prefix
         """
         self.logger = logging.getLogger(self.__class__.__module__ + "." + self.__class__.__name__)
 
-        self.redis = redis
+        self.redis_manager = redis_manager
         self.prefix = prefix
 
     def ping(self):
-        self.redis.ping()
+        self.redis_manager.ping()
 
     def set(self, key: str, value, ttl=600, pickling=False):
         """Set a cache item
@@ -42,7 +42,7 @@ class CacheClient(object):
             cachevalue = jsonDumps({"pickled": pickled})
         else:
             cachevalue = jsonDumps({"data": value})
-        self.redis.setex(self.prefix + key, ttl, cachevalue)
+        self.redis_manager.setex(self.prefix + key, ttl, cachevalue)
         self.logger.debug("Set cache item %s:%s [%ss]" % (key, truncate(cachevalue), ttl))
         return True
 
@@ -52,7 +52,7 @@ class CacheClient(object):
         :param key: cache item key
         :return: value
         """
-        value = self.redis.get(self.prefix + key)
+        value = self.redis_manager.get(self.prefix + key)
         if value is not None:
             # if found the cached data foe key
             # get envelop we axpect data for json-marshaled or pickeled for pickled
@@ -72,9 +72,9 @@ class CacheClient(object):
         :param ttl: item time to live [default=600s]
         :return: True
         """
-        self.redis.expire(self.prefix + key, ttl)
-        self.logger.debug("Set cache item %s expire to %s" % (key, ttl))
-        return True
+        ret = self.redis_manager.expire(self.prefix + key, ttl)
+        self.logger.debug("Set cache item %s expire to %s - ret: %s" % (key, ttl, ret))
+        return ret
 
     def delete(self, key):
         """Delete a cache item
@@ -82,7 +82,7 @@ class CacheClient(object):
         :param key: cache item key
         :return: True
         """
-        self.redis.delete(self.prefix + key)
+        self.redis_manager.delete(self.prefix + key)
         self.logger.debug("Delete cache item %s" % key)
         return True
 
@@ -92,10 +92,13 @@ class CacheClient(object):
         :param pattern: key search pattern
         :return: list of items
         """
-        keys = self.redis.keys(self.prefix + pattern)
+        # print("+++++ CacheClient - get_by_pattern - pattern: %s" % pattern)
+        keys = self.redis_manager.keys(self.prefix + pattern)
+        # print("+++++ CacheClient - get_by_pattern - keys: %s" % keys)
         res = []
-        for key in keys:
-            res.append({"key": key, "value": self.redis.get(key)})
+        if keys is not None:
+            for key in keys:
+                res.append({"key": key, "value": self.redis_manager.get(key)})
         return res
 
     def delete_by_pattern(self, pattern):
@@ -104,11 +107,9 @@ class CacheClient(object):
         :param pattern: key search pattern
         :return: True
         """
-        keys = self.redis.keys(self.prefix + pattern)
-        if len(keys) > 0:
-            res = self.redis.delete(*keys)
-            return res
-        return True
+        # print("+++++ CacheClient - delete_by_pattern - pattern: %s" % pattern)
+        res = self.redis_manager.delete(self.prefix + pattern)
+        return res
 
     def extend_ttl(self, key, ttl=600) -> bool:
         """Extend a cache item ttl
@@ -116,6 +117,6 @@ class CacheClient(object):
         :param key: cache item key
         :return: True
         """
-        self.redis.expire(self.prefix + key, ttl)
-        self.logger.debug("Extend cache item %s ttl to %s" % (key, ttl))
-        return True
+        ret = self.redis_manager.expire(self.prefix + key, ttl)
+        self.logger.debug("Extend cache item %s ttl to %s - ret: %s" % (key, ttl, ret))
+        return ret