+++ beecell/db/manager.py
@@ -404,9 +404,12 @@ class RedisManager(ConnectionManager):
         :param pattern: key search pattern [default='*']
         :return: list of tuple (key, type, ttl)
         """
+        # print("+++++ RedisManager - delete - pattern: %s" % pattern)
         keys = self.conn.keys(pattern)
+        # print("+++++ RedisManager - delete - keys: %s" % keys)
         if len(keys) > 0:
             res = self.conn.delete(*keys)
+            # print("+++++ RedisManager - delete - res: %s" % res)
             return res
         return None
 
@@ -545,6 +548,13 @@ class RedisManager(ConnectionManager):
         :return: lists of keys with value
         """
         return self.conn.lrem(*args, **kwargs)
+    
+    # new methods
+    def expire(self, name, time: int):
+        return self.conn.expire(name, time)
+
+    def keys(self, pattern):
+        return self.conn.keys(pattern)
 
 
 def manage_connection(method):